fix bugs

+++ loss.py
@@ -44,7 +44,7 @@ class CustomBCELoss(nn.Module):
     
     def forward(self,label_true, label_pred):
         label_mask = label_true.ne(-1)
-        label_loss = self.bce(label_pred[label_mask], label_true[label_mask].type(torch.LongTensor))
+        label_loss = self.bce(label_pred[label_mask].to(torch.long), label_true[label_mask].to(torch.float32))
         return label_loss
